# specs/003-complete-builder/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: PointApp Builder API
  version: 1.0.0
  license:
    name: Private and proprietary
    identifier: LicenseRef-Proprietary
  description: >-
    Authenticated Builder control-plane and public signed PointApp content-plane contract.
    Browser UI visibility never grants authority; every protected operation is authorized server-side.
servers:
  - url: https://appbuilder.pointatx.org
    description: Production
...
paths:
...
  /api/releases/validate:
    post:
      tags: [Releases]
      summary: Validate a saved revision
      operationId: validateRevision
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [revisionId]
              additionalProperties: false
              properties: { revisionId: { $ref: '#/components/schemas/Id' } }
      responses:
        '200':
          {
            description: Deterministic validation report.,
            content:
              { application/json: { schema: { $ref: '#/components/schemas/ValidationReport' } } },
          }
        '403': { $ref: '#/components/responses/Problem' }
  /api/releases/staging:
    post:
      tags: [Releases]
      summary: Publish a revision to PointApp Staging
      operationId: publishStaging
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [revisionId]
              additionalProperties: false
              properties: { revisionId: { $ref: '#/components/schemas/Id' } }
      responses:
        '201':
          {
            description: Signed immutable Staging release.,
            content: { application/json: { schema: { $ref: '#/components/schemas/ReleaseView' } } },
          }
        '400': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
  /api/releases/production/promotions:
    post:
      tags: [Releases]
      summary: Promote current Staging to Production
      operationId: promoteProduction
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [stagingReleaseId]
              additionalProperties: false
              properties: { stagingReleaseId: { $ref: '#/components/schemas/Id' } }
      responses:
        '201':
          {
            description: Production points to the exact current Staging release.,
            content: { application/json: { schema: { $ref: '#/components/schemas/ReleaseView' } } },
          }
        '403': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
  /api/releases/production/rollbacks:
    post:
      tags: [Releases]
      summary: Roll Production back to a prior release
      operationId: rollbackProduction
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [releaseId, reason]
              additionalProperties: false
              properties:
                releaseId: { $ref: '#/components/schemas/Id' }
                reason: { type: string, minLength: 8, maxLength: 240 }
      responses:
        '201':
          {
            description: Production points to a prior signed Production release.,
            content: { application/json: { schema: { $ref: '#/components/schemas/ReleaseView' } } },
          }
        '403': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
  /api/releases:
    get:
      tags: [Releases]
      summary: List release and channel history
      operationId: listReleases
      security: [{ sessionCookie: [] }]
      parameters:
        [{ $ref: '#/components/parameters/Cursor' }, { $ref: '#/components/parameters/Limit' }]
      responses:
        '200':
          description: Release and channel-event history.
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: '#/components/schemas/Page' }
                  - type: object
                    required: [items]
                    properties:
                      {
                        items: { type: array, items: { $ref: '#/components/schemas/ReleaseView' } },
                      }
        '403': { $ref: '#/components/responses/Problem' }
...
components:
  securitySchemes:
    sessionCookie: { type: apiKey, in: cookie, name: __Host-pointapp_builder_session }
  parameters:
    Cursor: { in: query, name: cursor, schema: { type: string, maxLength: 256 } }
    Limit:
      { in: query, name: limit, schema: { type: integer, minimum: 1, maximum: 100, default: 25 } }
    IdempotencyKey:
      { in: header, name: Idempotency-Key, required: true, schema: { type: string, format: uuid } }
...
  responses:
    Problem:
      description: Safe typed error.
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
...
  schemas:
    Id: { type: string, format: uuid }
...
    ValidationIssue:
      type: object
      required: [code, path, message, severity]
      properties:
        code: { type: string }
        path: { type: string }
        message: { type: string }
        severity: { type: string, enum: [error, warning] }
    ValidationReport:
      type: object
      required: [valid, revisionId, manifestDigest, reportDigest, issues]
      properties:
        valid: { type: boolean }
        revisionId: { $ref: '#/components/schemas/Id' }
        manifestDigest: { type: string, pattern: '^[a-f0-9]{64}$' }
        reportDigest: { type: string, pattern: '^[a-f0-9]{64}$' }
        issues: { type: array, items: { $ref: '#/components/schemas/ValidationIssue' } }
    ReleaseView:
      type: object
      required:
        [eventId, channel, eventKind, releaseId, revisionId, manifestDigest, keyId, createdAt]
      properties:
        eventId: { $ref: '#/components/schemas/Id' }
        channel: { type: string, enum: [staging, production] }
        eventKind: { type: string, enum: [publish, promote, rollback] }
        releaseId: { $ref: '#/components/schemas/Id' }
        revisionId: { $ref: '#/components/schemas/Id' }
        manifestDigest: { type: string, pattern: '^[a-f0-9]{64}$' }
        keyId: { type: string }
        signature: { type: string, contentEncoding: base64url }
        createdAt: { type: string, format: date-time }
...
    Page:
      type: object
      required: [items]
      properties: { nextCursor: { type: [string, 'null'] } }
    Problem:
      type: object
      required: [type, title, status, code, requestId]
      additionalProperties: false
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer, minimum: 400, maximum: 599 }
        code: { type: string, pattern: '^[A-Z0-9_]+$' }
        detail: { type: string }
        requestId: { $ref: '#/components/schemas/Id' }
        fields: { type: object, additionalProperties: { type: array, items: { type: string } } }